Moving hooks to different files
